Stop printing out messages when parsed output is empty.
This happens sometimes because our regex parsing is incomplete. I believe it comes out empty when packwerk is successful, for example.

In the future, if packwerk can output JSON, we can be more particular about what we expect. For now, we stop showing this error message so the user is not bothered or confused.